+++ src/Language/AST/Node.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace GraphQL\Language\AST;
 
 use GraphQL\Utils\Utils;
+use JsonSerializable;
 
 use function count;
 use function get_object_vars;
@@ -36,7 +37,7 @@ use function json_encode;
  * | ListTypeNode
  * | NonNullTypeNode
  */
-abstract class Node
+abstract class Node implements JsonSerializable
 {
     public ?Location $loc = null;
 
@@ -92,36 +93,29 @@ abstract class Node
 
     public function __toString(): string
     {
-        $tmp = $this->toArray(true);
-
-        return (string) json_encode($tmp);
+        return json_encode($this);
     }
 
     /**
      * @return array<string, mixed>
      */
-    public function toArray(bool $recursive = false): array
+    public function jsonSerialize(): array
     {
-        if ($recursive) {
-            return $this->recursiveToArray($this);
-        }
-
-        $tmp = (array) $this;
-
-        if ($this->loc !== null) {
-            $tmp['loc'] = [
-                'start' => $this->loc->start,
-                'end'   => $this->loc->end,
-            ];
-        }
-
-        return $tmp;
+        return $this->toArray();
     }
 
     /**
      * @return array<string, mixed>
      */
-    private function recursiveToArray(Node $node): array
+    public function toArray(): array
+    {
+        return self::recursiveToArray($this);
+    }
+
+    /**
+     * @return array<string, mixed>
+     */
+    private static function recursiveToArray(Node $node): array
     {
         $result = [
             'kind' => $node->kind,
@@ -147,11 +141,11 @@ abstract class Node
                 $converted = [];
                 foreach ($propValue as $item) {
                     $converted[] = $item instanceof Node
-                        ? $this->recursiveToArray($item)
+                        ? self::recursiveToArray($item)
                         : (array) $item;
                 }
             } elseif ($propValue instanceof Node) {
-                $converted = $this->recursiveToArray($propValue);
+                $converted = self::recursiveToArray($propValue);
             } elseif (is_scalar($propValue)) {
                 $converted = $propValue;
             } else {