+++ src/Utils/SchemaPrinter.php
@@ -40,11 +40,7 @@ use function strlen;
 /**
  * Prints the contents of a Schema in schema definition language.
  *
- * @phpstan-type Options array{commentDescriptions?: bool}
- *
- *  - commentDescriptions:
- *      Provide true to use preceding comments as the description.
- *      This option is provided to ease adoption and will be removed in v16.
+ * @phpstan-type Options array{}
  */
 class SchemaPrinter
 {
@@ -217,10 +213,6 @@ class SchemaPrinter
         $description = $def->description;
         if ($description === null) {
             return '';
-        }
-
-        if (isset($options['commentDescriptions'])) {
-            return static::printDescriptionWithComments($description, $indentation, $firstInBlock);
         }
 
         $preferMultipleLines = mb_strlen($description) > 70;