+++ tests/Language/VisitorTest.php
@@ -34,9 +34,7 @@ use function count;
 use function file_get_contents;
 use function func_get_args;
 use function gettype;
-use function is_array;
 use function is_numeric;
-use function iterator_to_array;
 
 class VisitorTest extends ValidatorTestCase
 {
@@ -113,16 +111,13 @@ class VisitorTest extends ValidatorTestCase
         ]);
     }
 
-    private function checkVisitorFnArgs($ast, $args, $isEdited = false): void
+    /**
+     * @param array<int, mixed> $args
+     */
+    private function checkVisitorFnArgs(DocumentNode $ast, array $args, bool $isEdited = false): void
     {
+        self::assertCount(5, $args);
         [$node, $key, $parent, $path, $ancestors] = $args;
-
-        $parentArray = $parent && ! is_array($parent)
-            ? ($parent instanceof NodeList
-                ? iterator_to_array($parent)
-                : $parent->toArray()
-            )
-            : $parent;
 
         self::assertInstanceOf(Node::class, $node);
         self::assertContains($node->kind, array_keys(NodeKind::CLASS_MAP));
@@ -141,8 +136,13 @@ class VisitorTest extends ValidatorTestCase
         }
 
         self::assertContains(gettype($key), ['integer', 'string']);
+        /** @var int|string $key */
 
-        self::assertArrayHasKey($key, $parentArray);
+        if ($parent instanceof NodeList) {
+            self::assertArrayHasKey($key, $parent);
+        } else {
+            self::assertObjectHasAttribute($key, $parent);
+        }
 
         self::assertIsArray($path);
         self::assertEquals($key, $path[count($path) - 1]);
@@ -154,7 +154,12 @@ class VisitorTest extends ValidatorTestCase
             return;
         }
 
-        self::assertEquals($node, $parentArray[$key]);
+        if ($parent instanceof NodeList) {
+            self::assertEquals($node, $parent[$key]);
+        } else {
+            self::assertEquals($node, $parent->{$key});
+        }
+
         self::assertEquals($node, $this->getNodeByPath($ast, $path));
         $ancestorsLength = count($ancestors);
         for ($i = 0; $i < $ancestorsLength; ++$i) {
@@ -163,15 +168,21 @@ class VisitorTest extends ValidatorTestCase
         }
     }
 
-    private function getNodeByPath(DocumentNode $ast, $path)
+    /**
+     * @param array<string|int> $path
+     *
+     * @return Node|NodeList
+     */
+    private function getNodeByPath(DocumentNode $ast, array $path): object
     {
         $result = $ast;
+
         foreach ($path as $key) {
-            $resultArray = $result instanceof NodeList
-                ? iterator_to_array($result)
-                : $result->toArray();
-            self::assertArrayHasKey($key, $resultArray);
-            $result = $resultArray[$key];
+            if ($result instanceof NodeList) {
+                $result = $result[$key];
+            } else {
+                $result = $result->{$key};
+            }
         }
 
         return $result;