+++ tests/Type/ValidationTest.php
@@ -1150,7 +1150,7 @@ class ValidationTest extends TestCase
     /**
      * DESCRIBE: Type System: Objects can only implement unique interfaces
      */
-    private function schemaWithObjectFieldOfType($fieldType): Schema
+    private function schemaWithObjectFieldOfType(Type $fieldType): Schema
     {
         $BadObjectType = new ObjectType([
             'name'   => 'BadObject',
@@ -1466,7 +1466,7 @@ class ValidationTest extends TestCase
         }
     }
 
-    private function schemaWithInterfaceFieldOfType($fieldType): Schema
+    private function schemaWithInterfaceFieldOfType(Type $fieldType): Schema
     {
         $BadInterfaceType = new InterfaceType([
             'name'   => 'BadInterface',
@@ -1582,7 +1582,7 @@ class ValidationTest extends TestCase
         }
     }
 
-    private function schemaWithArgOfType($argType): Schema
+    private function schemaWithArgOfType(Type $argType): Schema
     {
         $BadObjectType = new ObjectType([
             'name'   => 'BadObject',
@@ -1661,7 +1661,7 @@ class ValidationTest extends TestCase
         }
     }
 
-    private function schemaWithInputFieldOfType($inputFieldType)
+    private function schemaWithInputFieldOfType(Type $inputFieldType): Schema
     {
         $badInputObjectType = new InputObjectType([
             'name'   => 'BadInputObject',