+++ tests/Utils/SchemaExtenderTest.php
@@ -305,43 +305,6 @@ class SchemaExtenderTest extends TestCase
     }
 
     /**
-     * @see it('can describe the extended fields with legacy comments')
-     */
-    public function testCanDescribeTheExtendedFieldsWithLegacyComments(): void
-    {
-        $extendedSchema = $this->extendTestSchema('
-            extend type Query {
-                # New field description.
-                newField: String
-            }
-        ', ['commentDescriptions' => true]);
-
-        self::assertEquals(
-            $extendedSchema->getQueryType()->getField('newField')->description,
-            'New field description.'
-        );
-    }
-
-    /**
-     * @see it('describes extended fields with strings when present')
-     */
-    public function testDescribesExtendedFieldsWithStringsWhenPresent(): void
-    {
-        $extendedSchema = $this->extendTestSchema('
-            extend type Query {
-                # New field description.
-                "Actually use this description."
-                newField: String
-            }
-        ', ['commentDescriptions' => true]);
-
-        self::assertEquals(
-            $extendedSchema->getQueryType()->getField('newField')->description,
-            'Actually use this description.'
-        );
-    }
-
-    /**
      * @see it('extends objects by adding new fields')
      */
     public function testExtendsObjectsByAddingNewFields(): void
@@ -1297,23 +1260,6 @@ class SchemaExtenderTest extends TestCase
           """
           directive @new on QUERY
         ');
-
-        $newDirective = $extendedSchema->getDirective('new');
-        self::assertEquals('new directive', $newDirective->description);
-    }
-
-    /**
-     * @see it('sets correct description using legacy comments')
-     */
-    public function testSetsCorrectDescriptionUsingLegacyComments(): void
-    {
-        $extendedSchema = $this->extendTestSchema(
-            '
-          # new directive
-          directive @new on QUERY
-        ',
-            ['commentDescriptions' => true]
-        );
 
         $newDirective = $extendedSchema->getDirective('new');
         self::assertEquals('new directive', $newDirective->description);