+++ CHANGELOG.md
@@ -28,6 +28,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Call previously unused methods `EnumType::parseValue()` and `EnumType::parseLiteral()`
 - Strongly type `PromiseAdapter::createRejected()` to require `\Throwable`
 - Move members specific to `NamedType` out of `Type`: `$name`, `$description`, `$config`, `isBuiltInType()`, `assertValid()`
+- Always convert recursively when calling `Node::toArray()`
 
 ### Added
 
@@ -40,6 +41,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Add validation rule `UniqueEnumValueNames`
 - Add SDL validation rule `UniqueOperationTypes` (#995)
 - Add ability to remove custom validation rules after adding them via `DocumentValidator::removeRule()`
+- Make `Node` implement `JsonSerializable`
 
 ### Optimized
 
@@ -89,6 +91,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Remove `WrappingType::getWrappedType()` argument `$recurse` in favor of `WrappingType::getInnermostType()`
 - Remove `Type::assertType()`
 - Remove `ListOfType::$ofType`, `ListOfType::getOfType()` and `NonNull::getOfType()`
+- Remove option `commentDescriptions` from `BuildSchema::buildAST()`, `BuildSchema::build()` and `Printer::doPrint()`
+- Remove parameter `$options` from `ASTDefinitionBuilder`
 
 ## 14.11.3